softgpu: Move draw pixel code to separate file.
This separates things better anyway.  No major perf impact.
